fix: changed id type to Long on repositorys

+++ src/main/java/com/pedrohlopes/musicPub/repository/IArtistProfileRepository.java
@@ -3,5 +3,5 @@ package com.pedrohlopes.musicPub.repository;
 import com.pedrohlopes.musicPub.model.artist.ArtistProfileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IArtistProfileRepository extends JpaRepository<ArtistProfileEntity, Integer> {
+public interface IArtistProfileRepository extends JpaRepository<ArtistProfileEntity, Long> {
 }
