+++ src/main/java/com/pedrohlopes/musicPub/repository/IEstablishmentProfileRepository.java
@@ -3,5 +3,5 @@ package com.pedrohlopes.musicPub.repository;
 import com.pedrohlopes.musicPub.model.establishment.EstablishmentProfileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IEstablishmentProfileRepository extends JpaRepository<EstablishmentProfileEntity, Integer> {
+public interface IEstablishmentProfileRepository extends JpaRepository<EstablishmentProfileEntity, Long> {
 }
