+++ src/main/java/com/pedrohlopes/musicPub/repository/IUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface IUserRepository extends JpaRepository<UserEntity, Integer> {
+public interface IUserRepository extends JpaRepository<UserEntity, Long> {
 
     @Query(
             value = """
